pyproject.toml
# Airflow Mastery - Complete Apache Airflow 3.x Curriculum
# https://github.com/YOUR_ORG/airflow-mastery

[project]
name = "airflow-mastery"
version = "1.2.0"
description = "Complete Apache Airflow 3.x curriculum with hands-on exercises"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Airflow Mastery Team" }]
keywords = ["airflow", "data-engineering", "workflow", "orchestration", "education"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: System :: Distributed Computing",
]

# Core dependencies for running examples
dependencies = [
    "apache-airflow>=3.0.0,<4.0.0",
    "apache-airflow-providers-standard>=1.0.0",
    "apache-airflow-providers-cncf-kubernetes>=8.0.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "pendulum>=3.0.0",
]

[project.urls]
Documentation = "https://YOUR_ORG.github.io/airflow-mastery"
Repository = "https://github.com/YOUR_ORG/airflow-mastery"
Changelog = "https://github.com/YOUR_ORG/airflow-mastery/blob/main/CHANGELOG.md"

# ============================================
# Dependency Groups (uv sync --group <name>)
# ============================================

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocs-awesome-pages-plugin>=2.9.0",
    "mkdocs-minify-plugin>=0.8.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

# ============================================
# Tool Configuration
# ============================================

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".ruff_cache",
    "infrastructure",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "AIR",    # Airflow-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "S101",   # Use of assert (OK in tests)
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["dags", "plugins", "tests"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]
"modules/**/solutions/**/*.py" = ["S101", "T20"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "-ra"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["dags", "modules"]
omit = ["tests/*", "*/__pycache__/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
show_missing = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dags", "modules"]
